Feature/wifi (#54)
* Initial non blocking wifi server

* Implement with wifi controller

* Pre full integration test

* Integrate wifi into smart fuse box, only has relay at present, the remainder will form part of other PRs

+++ BoatControlPanel/ConfigCommandHandler.cpp
@@ -6,7 +6,7 @@ ConfigCommandHandler::ConfigCommandHandler(BroadcastManager* broadcastManager, H
 {
 }
 
-bool ConfigCommandHandler::handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], int paramCount)
+bool ConfigCommandHandler::handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], uint8_t paramCount)
 {
     // Access the in-memory config
     Config* cfg = ConfigManager::getConfigPtr();
