+++ Shared/CommandHandlers/InterceptDebugCommandHandler.h
@@ -19,7 +19,7 @@ public:
         return true;
     }
 
-    bool handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], int paramCount) override
+    bool handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], uint8_t paramCount) override
     {
 		(void)command;
 		(void)params;
